+++ src/main/java/chess/domain/game/PieceSettings.java
@@ -1,12 +1,6 @@
-package chess.domain;
+package chess.domain.game;
 
-import chess.domain.piece.Bishop;
-import chess.domain.piece.King;
-import chess.domain.piece.Knight;
-import chess.domain.piece.Pawn;
-import chess.domain.piece.Piece;
-import chess.domain.piece.Queen;
-import chess.domain.piece.Rook;
+import chess.domain.piece.*;
 
 public enum PieceSettings {
