+++ src/main/java/chess/domain/piece/Piece.java
@@ -1,7 +1,6 @@
 package chess.domain.piece;
 
-import chess.domain.Position;
-import chess.domain.Team;
+import chess.domain.game.Position;
 
 import java.util.ArrayList;
 import java.util.List;
