+++ src/main/java/chess/domain/piece/Queen.java
@@ -1,9 +1,8 @@
 package chess.domain.piece;
 
-import chess.domain.Position;
-import chess.domain.Team;
+import chess.domain.game.Position;
 
-import static chess.domain.MoveStrategy.QUEEN;
+import static chess.domain.piece.MoveStrategy.QUEEN;
 
 public class Queen extends Piece {
