+++ src/test/java/chess/domain/game/PositionTest.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.game;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
