+++ src/test/java/chess/domain/piece/QueenTest.java
@@ -3,8 +3,8 @@ package chess.domain.piece;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import chess.domain.Position;
-import chess.domain.Team;
+import chess.domain.game.Position;
+
 import java.util.List;
 import java.util.stream.Stream;
